[FIRRTL] Correct, Document Invalid Behavior (#1499)
Change FIRRTL dialect folders to use the Scala FIRRTL Compiler (SFC)
interpretation of invalid.  An invalid involving a non-mux primitive
operation is treated as a constant zero.  An invalid involving the true
or false paths of a multiplexer is treated as a conditionally valid
expression which always evaluates to the non-invalid path.

This intentionally deviates from the FIRRTL spec interpretation of
invalid as undefined behavior.  Invalidations are frequently generated
by Chisel (as a means of telling the SFC that an uninitialized component
is okay) and by a legacy version of Chisel ("Chisel 2 Compatibility
Mode") which defaults all connections to invalid to permanently disable
uninitialized checking.

Concretely, this adds the following boolean folds:

  - or(a, invalid) -> a
  - or(invalid, a) -> a
  - xor(a, invalid) -> a
  - xor(invalid, a) -> a

The following boolean folds already existed, but are documented with
information about the SFC interpretation of invalid:

  - and(a, invalid) -> 0
  - and(invalid, a) -> 0

Invalid value folders are also added for the following arithmetic
operations:

  - add(x, invalid) -> bits(x, width)
  - sub(x, invalid) -> bits(x, width)
  - mul(x, invalid) -> 0
  - div(invalid, x) -> 0
  - rem(invalid, x) -> 0

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>