[HIR] WIP MemrefLoweringPass. Addr bus now a tuple.
Address bus for memref is now a tuple of IntegerType so that individual
addresses can be assigned to individual elements of the tuple.

This helps in lowering LoadOp/StoreOp as well where muliple addr ssa
vars are packed into one tuple and sent to the bus.

This also would help in debugging if the code generator does not flatten
the addresses into one addr.