[FIRRTL] Enhance CHIRRTL memory ops and types
This change has several enhancements to the CHIRRTL memory operations
and types.

The two CHIRRTL memory operations have been renamed from `smem` to
`seqmem` and `cmem` to `combmem`. In addition, instead of returning
`!firrtl.vector<>` types, they return a new type  `!firrtl.cmemory`.
This new type can only be used with CHIRRTL memories and ports, and
prevents some shenanigans where it could be used like a normal vector.

There is a problem in SFC where memports could be used outside of the
scope of where they were declared in.  This is an issue with the pass
and verifier ordering, and not technically valid IR. Unfortunately this
leniency was abused in the Chisel standard library. After some talking
with SFC devs, we agreed to fix the problem in the standard library and
fix the hole in the verifier.  To allow MFC to consume FIRRTL with this
problem, there was a wire inserted into the memory port's parent scope
during parsing.  With a solid resolution of this problem in sight, this
logic is being remove from the parser.  For more information see
https://github.com/chipsalliance/chisel3/pull/2064.

The SeqMem, CombMem, and MemoryPort operations are all using the
ImplicitSSAName printer, which simplified the printing of the operation
when the `name ` attribute matched the SSA name.

The MemoryPortOp now implements the InferTypeOpInterface, which
simplified the type printing.