[FIRRTL] Fix type comparison bug for Bundles
Fix an issue where type equivalency was not existing after verifying
that two bundles were equivalent.  This resulted in a fall-through check
that the widthless type of two bundles _also_ being compared.  This
would cause issues where a bundle with an abstract reset would fail to
type check against a uint<1>.

Add a small test of this behavior.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>