[MSFT] Instance specific placement (#1550)
Placement data can now be filtered to a particular instance. Adds the SwitchInstance attribute to select an attribute from a list keyed off the instance ID (currently just the path in the instance hierarchy). This method is not robust to hierarchy changes, but it's a start.

Modified ExportTcl to respect the SwitchInstance attribute