[Comb] Fold a large series of of mux ops into array_create/get.
This substantially shrinks the size of the IR and generates much
better verilog.  E.g. before:

  module top_mod(
    input  [1:0] sel,
    input  [7:0] a, b, c, d,
    output [7:0] y);

    assign y = sel == 2'h0 ? a : sel == 2'h1 ? b : sel == 2'h2 ? c : d;
  endmodule

and after:

  module top_mod(
    input  [1:0] sel,
    input  [7:0] a, b, c, d,
    output [7:0] y);

    assign y = ({{a}, {b}, {c}, {d}})[sel];
  endmodule

This uses a couple heuristics to keep things from going crazy. This
is progress towards closing Issue #675, but it is still not getting
some important cases due to Issue #1549.  Those can be handled as
follow-up patches.