[CI] Run Windows build on push to main
This way, the dev who pushed (or merged the PR) gets a notification if they
broke the build. Also makes it easier to figure out which commit broke the
build.

Doing this now since I'm might start caring about Windows builds soon. Running
it on push to main only (instead of also on PRs) since most code changes don't
break it, so it's kinda wasteful.