[Comb] Add shift by 0 folders to `shl`, `shru`, `shrs`
This adds shift by constant 0 folders to all 3 shift operations in the
comb dialect.

This is actually a bugfix. Before this change, we would hit an assert
trying to contruct a 0 width APInt in the following code:
```c++
auto zeros =
    rewriter.create<hw::ConstantOp>(op.getLoc(), APInt::getNullValue(shift));
```

This also reorganizes the recently added test case
`shift_to_extract_and_concat` to follow the style of the rest of the
file, which is to have a separate test for each operation.