[Calyx] Add a MemoryOp primitive.
This follows a similar approach to the RegisterOp. Rather than
defining different primitives for different numbers of dimensions like
the Rust compiler, this uses a single operation to represent any
n-dimensional memory.

This generality means we can't enumerate explicit return types.
Rather, the return type is a variadic list of signless integer types.
A helper build method is provided to support this from C++ code, and a
verifier ensures everything lines up.