[FIRRTL][ExpandWhens] Simplify muxing of connected values
Expand whens takes the following code:
```firrtl
when a:
  x <= 1
else
  x <= 2
```
And transforms it turns it into the (desirable) code:
```firrtl
x <= mux(a, 1, 2)
```

However for nested whens, it was being overly pessimistic in the
condition of the mux.

Expand whens takes the following code:
```firrtl
when a:
  when b:
    x <= 1
  else:
    x <= 2
else
  x <= 3
```

And transforms it into
```firrtl
x <= mux(a, mux(a & b, 1, 2), 3)
```

When it should transform it into:
```firrtl
x <= mux(a, mux(b, 1, 2), 3)
``

In the code change `condition` refers to the condition of the current
WhenOp , and `thenCondition` refers to the all parent `WhenOp`
conditions `and`ed together.  The `thenCondition` is intended to be used
for simulation constructs which are copied directly into the module's
body.

The following example shows how `thenCondition` is used:
```firrtl
when a:
   when b:
     printf(clock, UInt<1>("h1"), "hi")
```

Here, `thenCondition` is `a & b`:
```firrtl
printf(clock, a & b & 1, "hi")
```