[FIRRTL][Parser] Stop removing no-op nodes when parsing
When a node is used as the address (aka index) of a CHIRRTL synchronous
read-only memory port, the memory port is enabled at the declaration
location of the node op.  Nodes are being removed by the parser if they
don't carry any annotations. When the node op is removed by the parser,
the enable conditions of the memory change, and sometimes the memory
port is never enabled.

This change removes the small optimization from the FIRParser so that
the memory port enable can be properly inferred.  These node operations
will still be removed later on during canonicalization.