[FIRRTL][CHIRRTL] Add `memoryport.access` operations (#1539)
The two CHIRRTL memory operations have been renamed from `smem` to
`seqmem` and `cmem` to `combmem`. In addition, instead of returning
`!firrtl.vector<>` types, they return a new type  `!firrtl.cmemory`.
This new type can only be used with CHIRRTL memories and ports, and
prevents some shenanigans where it could be used like a normal vector.

Memory ports in SFC are allowed to be used outside of the scope which they are
defined in. To work around this issue, the memory port declaration was split
into two operations:

1. `firrtl.memoryport`: This operation is the declaration of the memory port,
   and it should be emitted into the body of the module.

2. `firrtl.memoryport.acccess`: This operations is emitted to the location of
   the original memory port declaration, and is used for enable inference.

For more information about these new operations, and why we added them, see the
changes to FIRRTLRational.md.

This include changes to the FIR parser to emit these new operations.