[PrettifyVerilog] Fix unary sinking to keep constants nearby.
This pass is duplicating unary ops in order to make verilog prettier,
but can result in the op being in a different block than the constant.
This causes the constant to get emitted as a weird local param, which
can even be unused in the generated verilog.  Fix this to duplicate
the constant operands as well.