[PrettifyVerilog] Sink single-use expressions down the region tree.
This allows the verilog printer to print them inline, for example we'd
turn the example from Issue #1601 into:

```
module test(
  input clock, a, a2, a3);

  wire _T = a2 | a3;
  always @(posedge clock) begin
    `ifndef SYNTHESIS
      if (`PRINTF_COND_ & a)
        $fwrite(32'h80000002, "thing");
      if (`STOP_COND_ & _T)
        $fatal;
    `endif
  end // always @(posedge)
endmodule
```

We now produce: `if (`STOP_COND_ & (a2 | a3))` with no temporary
wire.  This is a pretty huge improvement to lots of things.

There is a more general version of this as well, but it seems best
to split into two patches.