[Calyx] Add IR support for calyx::WhileOp (#1605)
* [Calyx] Add IR support for calyx::WhileOp

This commit adds support for the Calyx While control operation. The operation takes two arguments; a boolean port and a group name. The op has a single body region representing the groups to execute when the condition evaluates to true.

This is solely for IR support and does not include:
- CalyxEmitter support (unsupported for all control ops as of now)
- CompileControl control FSM generation