[FIRRTL] Rewrite Grand Central Views/Interfaces
Rewrite the Grand Central Views/Interface pass to work with the new,
recursive structure representing interfaces.

Add an MLIR Attribute hierarchy to represent this recursive structure.
Use this to factor out verification logic from business logic that
generates interfaces.

After this rewrite, the pass now works in the following way:

1. Extraction information is pulled out from the circuit along with
   annotations that describe the interfaces.

2. The full circuit is walked to build up information about companions,
   parents, and leaves that will be connected to the interface.

3. Interface annotations on the circuit are then recursively walked to
   generate interfaces, instantiate interfaces, and generate XMRs in a
   "mappings" file that drives the interfaces.

A CircuitNamespace utility is added to generate unique symbols for
interfaces and binds that need to be created.

Interfaces (and other collateral) are only extracted if an
ExtractGrandCentralAnnotation is present.  This is a deviation from the
SFC Grand Central pass where absence of that annotation will cause an
assert to fire.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>