[HIR] Added/modified ops related to loops.
Instead of separate hir.for.next_iter and hir.while.next_iter we now
have hir.next_iter for both for/while loops.

hir.condition is a separate op now. It does not have time associated
with it. The time of the condition var must be same as the hir.next_iter
op for the loop.

hir.is_first_iter is a new op which helps in converting ForOp to
WhileOp.