[Calyx] Add a primitive cell instance operation
This operation represents an instance of a circuit primitive. The operation function is specified by the `$func` parameter. The result types of the parameter must match the natural type constraints of the provided `$func`. I.e. for binary arithmetic operations the result and input types are expected to be the identical, whereas for boolean comparison operators input types should be identical with a 1-bit wide output type.

The available operations maps to the [native Calyx logical and primitive operations](https://github.com/cucapra/calyx/blob/master/primitives/core.futil), please let me know whether you think it'd be smarter to map to the `std.comb` operations instead (or a mix of the two).

I'd like to discuss whether we should constrain the "prim" operations to be purely combinational standard operations. This would imply that if we in the future add an operation for i.e., multi-cycle multiplication (which would have `clk,reset,go,done` signals), this would have to be a separate calyx component. If so, we might also want to change to operation name from `calyx.prim` to i.e. `calyx.comb`.