[Comb] Fix shift by constant canonicalizers (#1640)
The shift by constant canonicalizers extract the relevant bits from the
value and prepend/append 0s depending on the direction shifted. The
canonicalizer made a mistake in the extract op generation, chopping off
the high bits when shifting right, and chopping off the low bits when
shifting left. When doing a signed ShrS, it was using the low bit of the value
as the sign bit.