Non Local Annotations, Distributed Global Table hook (#1631)
Build a flattened tree for non-local annotations. This provides a global symbol for a NLA's path. Each component (instance and final annotation) in a non-local annotation carries a field pointing to the is global.
Non-local annotations are then identified by having a pointer to the global op in the circt.nonlocal field of an annotation.
Each element along the way carries an annotation with class circt.nonlocal which has a matching circt.nonlocal field pointing to the global op. Thus instances participating in non-local annotation paths are readily apparent.

The check shows the intended scattering, which is part of the a different patch.

The only places you need to know you are part of a NLA are at instance ops and at the final named thing. This will put
non-local annotations on the final named thing, as local annotations are, with a flag (pointer to a global path via this new op). It's not expected that anything much has to change in annotation consumers due to this.

The instances are what form paths, so they have a flag to say they are part of an NLA. There are two cases:

The final annotation is on the instances, in which case it will have an annotation with whatever class and the annotation will have a field 'circt.nonlocal' pointing to the global path op. In general, passes need to understand annotations on operations to manipulate them, so the normal check here is "does it have any annotations?" which is fast, catches both local and non-local.

The instance is part of a path. As before, a conservative "does it have any annotations" will suffice for most firrtl passes. If a pass is aware of some annotations, it is already walking the annotation list to make sure it understands all the annotations, in which case it will see an annotation with the class "circt.nonlocal" indicating a non-local annotation goes through the instance. I believe that any pass which manipulates an instances already either has to be aware of annotations, in which case this extra check is already necessary and part of what they are already doing, or only manipulates non-annotated instances, in which case the "is it empty" test is fast.