[ExportVerilog] Adopt hw::SymbolCache to speed things up, NFC.
This builds a SymbolCache at setup time, when the verilog exporter
is scanning the ops to see what files they go into.  This cache is
then used when emitting each of the chunks, speeding up the
"getReferencedFoo" calls by turning them into O(1) lookups.

This speeds up export verilog on a large testcase that includes a
bunch of binds from 126s to 17s.  It should also help speedup
other normal cases as well, since each instance has to resolve the
module it refers to.