[ExportVerilog] Add always_comb lowering option
This add an option to control how `always_comb` is printed by the
verilog exporter.  The default has been changed to print `always_comb`
as `always @(*)`. A lowering option `alwaysComb` has been added, similar
to `alwaysFF` to use the old printing style.  This change is motivated
by a lack of tool support for the `always_comb` verilog construct.