[ExportVerilog] Add verifLabels lowering option
Add the `verifLabels` lowering option which assigns an automatically
generated label to all verification statements (assert/assume/cover)
that have no label specified. This works around the issue of
verification EDA tools commonly requiring all verification statements to
be labeled.

This commit adds the labels of these ops to the name manager in the
ExportVerilog prepass, and then uses the manager to uniquify a generic
label generated for unlabeled ops in case the `verifLabels` option is
set.

Fixes #1691.