[ExportVerilog] Fix Issue #1681, miscompile of >>> operator.
It turns out that Verilog has some weird context sensitive sign inference
thing going on that affects the result type of the >>> operator and causes
it to do an unsigned shift.  I cannot pin down how or why this happens,
so we'll just emit redundant $signed() casts around the operator to make sure
it is controlled appropriately.

I don't feel very good about not understanding this from reading the spec,
but empirically this is how it works with verilog implementations.