[FIRRTL] Add dedicated PortDirectionsAttr
This adds a dedicated attribute to represent the `in` and `out`
directions of module ports.  This is represented under the hood using an
LLVM bitvector.

This replaces using an `IntegerAttr` as a bitvector for the port
directions.  This has a large limitation in that it cannot represent 0
ports, since it has a minimum bit-width of 1.  To make this work, the
old implementation relied on using the relevant Op to determine the
number of ports.  As such, this Attribute could not stand on its own.

Cleaning this up will make it easier to share the port direction
information stored in FModuleLike operations with the Instance Op.