[FIRRTL] Add port direction information to InstanceOp
This duplicates the port direction information from the module operation
on to the instance op.  This will let us query information about port
direction without looking up the related module.  This can be used to
speed up connect flow verification, and many passes.

The port direction is printed in the ir interleaved with the port types,
similar to modules:
```mlir
    %result = firrtl.instance @bar740 {name = "fpga"} : in !firrtl.analog<1>
```