[Calyx] Add emission for component Block Arguments. (#1726)
Before, there was an early short-circuit to ensure that if the Value has no defining operation, we would not 
TypeSwitch on it. A block argument will not have a defining operation. This led to the problem in #1724; 
since %done is a block argument of the ComponentOp, we just short-circuited. This pull request now 
checks to see if the value is a BlockArgument, and emits the respective port.