[StaticLogic] Add new PipelineWhile operation.
https://llvm.discourse.group/t/rfc-representing-pipelined-loops/4171/1
for more discussion.

This defines the operation's syntax more or less as discussed,
including round-trip tests for the four running examples we've had.

In order to reduce test churn, the existing PipelineOp and the
StandardToStaticLogic pass are left untouched for now. Ultimately, the
new pipeline operation could replace the existing PipelineOp.