[Calyx] Fix twine usage in emitter
Twines are not safe to store on the stack. This caused undefined
behaviour which lead to an issue where it would print `group` twice
depending on the compiler settings.

```
// Correct:
comb group Group2
// Incorrect:
comb group Group2comb group
```

This issue was noted on the forums here: https://llvm.discourse.group/t/calyx-test-failure/4261