[FIRRTL] GCT Views: End-to-end "Wire" Test, NFC
Add an end-to-end test of the Grand Central Views/Interfaces pass.  This
compiles a FIRRTL circuit with a non-trivial interface structure all the
way to Verilog.  This circuit is compiled with both extraction enabled
and disabled.

The circuit consists of a non-trivial aggregate wire composed of ground
types, vector types, bundle types, and vectors of bundle types.  This
wire is then passed from a DUT, into a submodule, drives a wire in the
submodule, and is then pased back out.  The interface is then
constructed from connections to the aggregate wire in the DUT and the
submodule.

For background, this is derived from a local test case built using the
Grand Central Chisel API.  The interface instantiation names are changed
to better exercise the Grand Central pass.  (Chisel Grand Central will
keep the interface name and interface name the same, but this can hide
bugs in the CIRCT Grand Central pass.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>