Update LLVM to 704a395 (#1765)
This updates LLVM to 704a395.  We are pulling the new version of LLVM for the top commit which allows 0-bit APInts.  This will help us model 0 bit constants in FIRRTL, and we will be able to properly use IntegerAttr as a (possibly 0-length) bitvector for FIRRTL port directions.

Included in this update is a change to the way `unrealized_conversion_casts` are handled by the LLVM converter infrastructure.  We needed to update `LLHDToLLVM` to apply the new conversion casts.  For more information see the public PSA on the LLVM discourse[^0].

This revealed a problem in the `LLHDToLLVM` where there was a failed to legalize operations in the lowering of `llhd.proc` and `llhd.entity`.  These operation's lowering logic needed to ensure that operations from the (old) body of `llhd.proc` and `llhd.entity` were type converted before inserting new LLVM operations which used their results.

[0] https://llvm.discourse.group/t/psa-run-reconcile-unrealized-casts-after-all-convert-to-llvm-from-now-on/4266
Co-authored-by: Martin Erhart <maerhart@outlook.com>