[FIRRTL][IMConstProp] Fix Unknown DontTouch ports
When a module port is marked DontTouch, we can't propagate a value
through it, and they should be marked as Overdefined.  They were
accidentally being recorded as Unknown, which caused them to behave the
same as Invalid value attributes.  This notably would cause incorrect
optimization when the value of a DontTouch output port was used as a
value in a mux: it would optimize the mux to the other value.

This change correctly marks all users of an output port as overdefined
when the output port is driven.

Output ports can be used as source values in their block, and need to be
marked as Overdefined. This is already correctly handled by
`mergeLatticeValue`, which will mark any value as overdefined if it is
marked DontTouch.