[Handshake] Postpone adding ctrl arg to handshake::FuncOp until end of conversion
No functional change, but a major step in the debug-ability of the pass. Previously, there was a discrepancy between the # of args of the Handshake FuncOp and the inlined body of the source FuncOp. This meant that any attempt at dump'ing the IR during conversion would crash, due a mismatch between the type signature of the handshake funcop and its entry block arguments.

This commit postpones adding the control argument to (and modifying the type signature of) the handshake funcOp until end of conversion.