[Handshake] Fix lowering of std.CallOp to handshake.InstanceOp (#1538)
The PR adds a wrapRewriter which wraps the partial lowering functions already in place (such as replaceCallOps, addForkOps...). This rewriter composes a bit of supporting code (lowerToHandshake, PartialLowerFuncOp) that creates an instance of a ConversionTarget as well as an application of applyPartialConversion. The end result is that any rewriting and their side effects (most importantly modifications to referencing SSA values) made within each of the partial lowering functions are fully "comitted" after a partial lowering function returns. While being a bit hacky, I suspect that this structure might be generally useful for conversion passes which is heavily dependent on some sequence of transformations being applied in a specific order, where the author does not want to write distinct transformation passes for each of these steps.