[SCFToCalyx] Add partial lowering infrastructure [4/13]
This commit adds a partial lowering infrastructure to the SCFToCalyx pass. Similarly to `StandardToHandshake`, `SCFToCalyx` structures its conversion into multiple distinct stages. Some of these stages perform IR modifications instead of IR replacements. The partial lowering infrastructure allows for writing passes which are driven by the conversion infrastructure, but relaxes the requirement of "if a conversion pattern matches, the matched operation should be replaced or erased".

`FuncOpPartialLoweringPattern`s are patterns which intend to match on `FuncOp`s and then perform their own walking from these. These patterns are initialized with access to the `ComponentLoweringState` for corresponding component of the matched `FuncOp`.