[SCFToCalyx] Build an execution schedule by CFG traversal [7/13]
This commit builds a control schedule by traversing the CFG of the function and associating this with the previously created groups. For simplicity, the generated control flow is expanded for all possible paths in the input DAG. This elaborated control flow will later be reduced in the various control schedule simplification patterns.