[FIRRTL][LowerToHW] Pass through `sv.verbatim` ops
The LowerToHW pass will emit an error for any unrecognized operation it
finds in the Circuit while lowering.  We need to pass through
`sv.verbatim` operations which are created during metadata generation.
The `sv.interface` operation is already passed directly through when
lowering, and this change extends this to include `sv.verbatim`
operations.  If we find that there are too many operations being
white-listed, we could modify it to pass-through anything by default.