[HW] Make OutputFileAttr a AttrDef instead of StructAttr
This moves OutputFileAttr from an AttrDef to a StructAttr.  This
StructAttr prevented us from puting member functions on the attribute
which were need.

This adds the ability to normalize filenames when creating an
OutputFileAttr, which is helpful when creating these from FIRRTL
annotations.

This also removes the cmake magic for generating StructAttrs in the HW
dialect, as there were none left.