[SCFToCalyx] Add support for building combinational logic [5/13] (#1830)
* [SCFToCalyx] Add support for building combinational logic [5/13]

This commit adds support for building `ComponentOp`'s and  most combinational operators from the `std` dialect. The combinational operators are built through the `buildLibOp` function. Here, we instantiate a new Calyx primitive, create a combinational group, and assign the input ports of the calyx primitive with the operands of the source operation.