[FIRRTL] Validate reg async reset value is constant at LowerToHW Time (#1444)
* [FIRRTL] Rename ResetType to AnyResetType, add actual ResetType

The FIRRTL dialect currently has things like `AsyncResetType` or
`ClockType` which are the actual types, but `ResetType` is a constraint
that captures `asyncreset`, `reset`, and `uint<1>`. Rename this to
`AnyResetType` and add a `ResetType` that represents the actual
buildable `firrtl.reset` type in analogy to the others.

* [FIRRTL] Add async reset const check to LowerToHW

When lowering a `firrtl.regreset` with async reset in `LowerToHW`, check
that the reset value is a constant. This is a common constraint imposed
by EDA software and is enforced on the Scala side as well (in the
CheckResets pass).

This also finally enables the `async-reset-errors.mlir` integration test
that we had for quite a while now.